Moved avatar to ProfileTab
